user_service/requirements.txt
asyncpg==0.29.0
bcrypt==4.2.0
fastapi==0.112.2
loguru==0.7.2
psycopg2-binary==2.9.9
pyhumps==3.8.0
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
uvicorn==0.30.6
